requirements.txt
streamlit==1.32.2
torch==2.2.0
transformers==4.38.2
sentence-transformers==2.5.1
torch-geometric==2.5.0
torch-scatter==2.1.2
torch-sparse==0.6.18
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1
umap-learn==0.5.5
plotly==5.19.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
faiss-cpu==1.8.0
rank-bm25==0.2.2
networkx==3.2.1
pyvis==0.3.2
spacy==3.7.4
scispacy==0.5.4
en-core-sci-md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_md-0.5.4.tar.gz
aiohttp==3.9.3
tqdm==4.66.2
openai==1.12.0